:art: Improve project structure by using CommonJS
